completed admin dahboard

+++ backend/Controllers/admin/sellerFunctions.js
@@ -6,27 +6,59 @@ const applySeller = require('../../models/seller/applySellerSchema');
 const sendEmail = require('../../utils/sendEmail');
 const handleError = require('../../utils/errorHandler');
 const generateCode = require('../../utils/generateCode');
+const authorizeChangeMiddleware = require('../../middleware/authorizeChangeMiddleware');
+const customLogger = require('../../utils/logHandler');
 
 // Get applied sellers
 const getAllApplySellers = async (req, res, next) => {
     try {
-        const applySellers = await applySeller.find().populate('user', 'fullName email'); // populate the user field with only fullName and email
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
+        const applySellers = await applySeller.find(); // populate the user field with only fullName and email
 
         res.status(200).json({
             success: true,
             message: 'All apply sellers retrieved successfully',
-            data: applySellers,
+            sellers: applySellers,
         });
     } catch (error) {
         return handleError(res, err);
     }
 };
 
+// get applied seller by id
+const getApplySellerById = async (req, res, next) => {
+    try {
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
+
+        const sellerId = req.params.sellerId;
+        const seller = await applySeller.findById(sellerId).populate({ path: 'user', select: '-tokens -orders -tickets -wishlist -cart -password' });
+
+        res.status(200).json({
+            success: true,
+            message: 'Seller retrieved successfully',
+            seller: seller,
+        });
+    } catch (error) {
+        return handleError(res, error);
+    }
+};
+
 // Accept seller
 const acceptSeller = async (req, res, next) => {
+    customLogger(req.user.role, "Aceept seller", req)
     try {
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
         // Find the seller application to be approved
-        const newSellerApplication = await applySeller.findById(req.params.id);
+        const newSellerApplication = await applySeller.findById(req.params.sellerId);
 
         if (!newSellerApplication) {
             return handleError(res, {
@@ -52,10 +84,11 @@ const acceptSeller = async (req, res, next) => {
         });
 
         // Save the new seller instance to the database
-        const savedSeller = await newSeller.save();
+        await newSeller.save();
 
         // Remove the seller application from the applySellerSchema
-        await newSellerApplication.remove();
+        newSellerApplication.status = "Approved";
+        await newSellerApplication.save();
 
         // Update user in the application
         const user = await User.findById(newSeller.user._id);
@@ -70,29 +103,78 @@ const acceptSeller = async (req, res, next) => {
         newSeller.verificationCodeExpiresAt = codeExpiry;
         await newSeller.save();
 
+        res.status(200).json({
+            status: "success",
+            message: 'Seller account created successfully'
+        });
+
         const data = {
             subject: 'New Seller Account - SwiftMarket',
             username: newSeller.businessName,
             verificationCode: newSeller.verificationCode,
-            verificationLink: 'https://example.com/verify'
+            verificationLink: `http://localhost:3001/verify/${newSeller.businessEmail}`
         };
 
         await sendEmail(newSeller.businessEmail, data, './verification/verifySeller.hbs');
 
-        res.status(201).json({
-            success: true,
-            message: 'Seller account created successfully',
-            data: savedSeller,
-            subject: 'New Seller Account - SwiftMarket'
-        });
     } catch (error) {
         return handleError(res, error);
     }
 };
 
+// reject seller
+const rejectSeller = async (req, res, next) => {
+    customLogger(req.user.role, "reject seller", req)
+    try {
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
+        // Find the seller application to be approved
+        const newSellerApplication = await applySeller.findById(req.params.sellerId);
+
+        if (!newSellerApplication) {
+            return handleError(res, {
+                code: 'not_found',
+                status: 'error',
+                message: 'Seller application not found',
+            });
+        }
+
+        // Remove the seller application from the applySellerSchema
+        newSellerApplication.status = "Rejected";
+        await newSellerApplication.save();
+
+        res.status(200).json({
+            status: "success",
+            message: 'Seller request rejected successfully',
+            subject: 'New Seller Account Request Rejected - SwiftMarket'
+        });
+
+        // Update user in the application
+        const user = await User.findById(newSellerApplication.user._id);
+
+        const data = {
+            subject: 'New Seller Account Rejected - SwiftMarket',
+            username: user.username,
+            reason: req.body.reason
+        };
+
+        await sendEmail(user.email, data, './seller/rejectSeller.hbs');
+
+    } catch (error) {
+        return handleError(res, error);
+    }
+}
+
 // Get All Verfied sellers
 const getAllSellers = async (req, res, next) => {
     try {
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
+
         const sellers = await Seller.find({}).populate('user', 'fullName email'); // populate the user field with only fullName and email
 
         res.status(200).json({
@@ -107,7 +189,14 @@ const getAllSellers = async (req, res, next) => {
 
 // Create a new seller
 const createSeller = async (req, res) => {
+    customLogger(req.user.role, "create seller", req)
     const errors = validationResult(req);
+
+    const result = authorizeChangeMiddleware("seller", req, res);
+    if (result === -1) {
+        return;
+    }
+
     if (!errors.isEmpty()) {
         return handleError(res, {
             code: 'CustomValidationError',
@@ -187,8 +276,29 @@ const createSeller = async (req, res) => {
 
 };
 
+// Get a single seller
+const getSellerById = async (req, res) => {
+    try {
+        const { id } = req.params;
+
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
+
+        const seller = await Seller.findById(id);
+        if (!seller) {
+            res.status(400).json({ status: 'error', message: "seller not found" });
+        }
+        res.status(200).json({ status: 'success', message: 'Seller retrieved', seller });
+    } catch (err) {
+        return handleError(res, err);
+    }
+};
+
 // Update seller
 const updateSeller = async (req, res) => {
+    customLogger(req.user.role, "update seller", req)
     try {
         const errors = validationResult(req);
         if (!errors.isEmpty()) {
@@ -199,8 +309,11 @@ const updateSeller = async (req, res) => {
             });
         }
 
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
         const { id } = req.params;
-        const updates = req.body;
 
         // check if seller exists
         const seller = await Seller.findById(id);
@@ -213,25 +326,51 @@ const updateSeller = async (req, res) => {
         }
 
         // update seller information
-        Object.keys(updates).forEach((key) => (seller[key] = updates[key]));
+        const {
+            businessUsername,
+            businessEmail,
+            businessRegistrationNumber,
+            businessNumber,
+            businessName,
+            businessLogo,
+            businessType,
+            businessWebsite,
+            taxIDNumber,
+            productCategories,
+            businessAddress,
+        } = req.body;
+
+        if (businessEmail) seller.businessEmail = businessEmail;
+        if (businessNumber) seller.businessNumber = businessNumber;
+        if (businessName) seller.businessName = businessName;
+        if (businessUsername) seller.businessUsername = businessUsername;
+        if (businessLogo) seller.businessLogo = businessLogo;
+        if (businessRegistrationNumber) seller.businessRegistrationNumber = businessRegistrationNumber;
+        if (businessType) seller.businessType = businessType;
+        if (businessAddress) seller.businessAddress = businessAddress;
+        if (businessWebsite) seller.businessWebsite = businessWebsite;
+        if (taxIDNumber) seller.taxIDNumber = taxIDNumber;
+        if (productCategories) seller.productCategories = productCategories;
+
         await seller.save();
+
+        res.status(200).json({ status: 'success', message: 'Seller updated successfully' });
 
         const data = {
             sellerUpdated: {
-                code: seller.businessName,
+                name: seller.businessName,
                 email: seller.businessEmail
             },
             violation: {
-                code: req.body.violationName,
+                name: req.body.violationName,
                 reason: req.body.violationReason,
-                adminUsername: req.user.fullname,
+                adminUsername: req.user.username,
             },
             subject: 'Seller Updated - SwiftMarket'
         };
 
         await sendEmail(seller.businessEmail, data, './violationOfTerms/sellerTerms.hbs');
 
-        res.status(200).json({ status: 'success', message: 'Seller updated successfully', data: seller });
     } catch (err) {
         return handleError(res, err);
     }
@@ -239,8 +378,13 @@ const updateSeller = async (req, res) => {
 
 // Delete seller
 const deleteSeller = async (req, res) => {
+    customLogger(req.user.role, "delete seller", req)
     try {
         const { id } = req.params;
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
         const seller = await Seller.findByIdAndDelete(id);
 
         if (!seller) {
@@ -250,26 +394,71 @@ const deleteSeller = async (req, res) => {
                 message: 'Seller not found',
             });
         }
+        res.status(200).json({ status: 'success', message: 'Seller deleted successfully' });
 
         const data = {
             sellerDeleted: {
-                code: seller.businessName,
+                name: seller.businessUsername,
                 email: seller.businessEmail
             },
             violation: {
-                code: req.body.violationName,
+                name: req.body.violationName,
                 reason: req.body.violationReason,
-                adminUsername: req.user.fullname,
+                adminUsername: req.user.username,
             },
             subject: 'Seller Account Deleted - SwiftMarket'
         };
 
         await sendEmail(seller.businessEmail, data, './violationOfTerms/sellerTerms.hbs');
 
-        res.status(200).json({ status: 'success', message: 'Seller deleted successfully' });
     } catch (err) {
         return handleError(res, err);
     }
 };
 
-module.exports = { getAllApplySellers, acceptSeller, getAllSellers, createSeller, updateSeller, deleteSeller }
+// ban seller
+const banSeller = async (req, res) => {
+    customLogger(req.user.role, "ban seller", req)
+    try {
+        const { id } = req.params;
+        const seller = await Seller.findById(id);
+
+        if (!seller) {
+            return handleError(res, {
+                code: 'not_found',
+                status: 'error',
+                message: 'Seller not found',
+            });
+        }
+
+        const result = authorizeChangeMiddleware("seller", req, res);
+        if (result === -1) {
+            return;
+        }
+
+        seller.banStatus.isBanned = true;
+        seller.banStatus.banExpiresAt = req.body.expiresAt;
+        await seller.save();
+
+        res.status(200).json({ status: 'success', message: 'Seller has been banned successfully' });
+
+        const data = {
+            userBanned: {
+                name: seller.businessUsername,
+                email: seller.businessEmail,
+            },
+            violation: {
+                name: req.body.violationName,
+                reason: req.body.violationReason,
+                adminUsername: req.user.username,
+            },
+            subject: 'Account banned - SwiftMarket'
+        };
+
+        await sendEmail(seller.businessEmail, data, './violationOfTerms/userTerms.hbs');
+    } catch (err) {
+        return handleError(res, err);
+    }
+};
+
+module.exports = { rejectSeller, getApplySellerById, banSeller, getSellerById, getAllApplySellers, acceptSeller, getAllSellers, createSeller, updateSeller, deleteSeller }
